teacher func upgate

+++ server/controllers/calendarController.js
@@ -16,10 +16,11 @@ exports.createCalendar = async (req, res) => {
 exports.listCalendar = async (req, res) => {
     try {
         const calendar = await Calendar.find({}).populate({
-            path:"coursee",
-            populate:{
-                path:"room"
-            }
+            path: "coursee",
+            populate: {
+                path: "room"
+            },
+
         })
         res.send(calendar);
     } catch (err) {
@@ -44,10 +45,26 @@ exports.updateCalendar = async (req, res) => {
 
 exports.deleteCalendar = async (req, res) => {
     try {
-        const calendar = await Calendar.findOneAndDelete(
-            { _id: req.params.id })
+
+        const course = await Calendar.findOne(
+            { _id: req.params.id }).populate("coursee").select("coursee")
+        if (req.user.role == 'admin') {
+            const calendar = await Calendar.findOneAndDelete(
+                { _id: req.params.id })
+            res.send(calendar);
+        } else {
+            if (req.user.user_id == course.coursee.teacher) {
+                const calendar = await Calendar.findOneAndDelete(
+                    { _id: req.params.id })
+                res.send(calendar);
+            } else {
+                res.status(400).send("you have no rights")
+            }
+        }
+
+
         // console.log(req.params.id)
-        res.send(calendar);
+        // res.send("calendar");
     } catch (err) {
         console.log(err);
         res.status(500).send("Server Error!!! on list calendar");
